requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astunparse==1.6.3
attrs==22.1.0
audioread==3.0.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
dataclasses==0.6
decorator==5.1.1
dill==0.3.5.1
dm-tree==0.1.7
etils==0.8.0
fire==0.4.0
flatbuffers==1.12
fonttools==4.37.4
gast==0.4.0
gin-config==0.5.0
google-api-core==2.8.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.3.3
google-cloud-bigquery-storage==2.16.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.48.2
grpcio-status==1.48.2
h5py==3.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
joblib==1.2.0
kaggle==1.5.12
keras==2.9.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labelImg==1.8.6
libclang==14.0.6
librosa==0.8.1
llvmlite==0.36.0
lml==0.1.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.3
neural-structured-learning==1.4.0
numba==0.53.0
numpy==1.20.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==20.9
pandas==1.5.0
Pillow==9.2.0
pooch==1.6.0
promise==2.3
proto-plus==1.22.1
protobuf==3.19.6
psutil==5.9.2
py-cpuinfo==8.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycocotools==2.0.5
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
sentencepiece==0.1.97
six==1.15.0
sounddevice==0.4.5
soundfile==0.11.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.18.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
tensorflow-model-optimization==0.7.3
tensorflowjs==3.18.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.3.4
tflite-support==0.4.1
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.1
typeguard==2.13.3
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.25.11
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.8.1
